Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Why would it hurt? You can't be using this as if the comments are true then the build fails.

There's no reason to deny this imo - until someone fixes it, this prevents someone walking in to a build error. It's not like it's a permanent decision.

Patch-set: 1